JACOBIN-102 Converted operand stack to int64 entries (vs. previous int32)

+++ src/exec/classes.go
@@ -16,10 +16,10 @@ import (
 // Classes contains all the loaded classes
 var Classes = make(map[string]Klass) // TODO: make these maps sync.Map
 
-// Statics is a fast-lookup map of static variables and functions. The int value
+// Statics is a fast-lookup map of static variables and functions. The int64 value
 // contains the index into the statics array where the entry is stored.
 // Statics are placed into this map only when they are first referenced and resolved.
-var Statics = make(map[string]int32)
+var Statics = make(map[string]int64)
 var StaticsArray []Static
 
 type Klass struct {
